server: move the loopback code to package `netutil`
Release note: None
