+++ pkg/util/netutil/BUILD.bazel
@@ -4,6 +4,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "netutil",
     srcs = [
+        "loopback.go",
         "net.go",
         "srv.go",
     ],
